[cleanup] Resolved a SonarCube bug

+++ src/main/java/com/github/hanyaeger/core/scenes/delegates/CoordinateGridDelegate.java
@@ -65,8 +65,8 @@ public class CoordinateGridDelegate implements Destroyable {
 
     private void drawCoordinateText(final int x, final int y, final int value, final int deltaX, final int deltaY) {
         var text = textFactory.createText(Integer.valueOf(value).toString(), TEXT_COLOR);
-        text.setLayoutX(x + deltaX);
-        text.setLayoutY(y + deltaY);
+        text.setLayoutX((double) x + deltaX);
+        text.setLayoutY((double) y + deltaY);
         pane.getChildren().add(text);
         gridElements.add(text);
     }
